@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
+apply plugin: "maven-publish"
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    namespace = "com.msinghal34.pinlockviewapp"
+    compileSdk 34
 
     defaultConfig {
-        applicationId "com.andrognito.pinlockviewapp"
-        minSdkVersion 11
-        targetSdkVersion 25
+        minSdkVersion 21
+        targetSdkVersion 34
         versionCode 1
         versionName "1.0"
     }
@@ -20,8 +20,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile project(':pinlockview')
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'junit:junit:4.13.2'
+    implementation "androidx.appcompat:appcompat:1.7.0"
+    implementation "androidx.recyclerview:recyclerview:1.3.2"
+    implementation project(':pinlockview')
 }
