@@ -1,54 +1,56 @@
 apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: "maven-publish"
 
-ext {
-    bintrayRepo = 'maven'
-    bintrayName = 'pin-lock-view'
-
-    publishedGroupId = 'com.andrognito.pinlockview'
-    libraryName = 'pinlockview'
-    artifact = 'pinlockview'
-
-    libraryDescription = 'A clean, minimalistic, easy-to-use and highly customizable pin lock custom view'
-
-    siteUrl = 'https://github.com/aritraroy/PinLockView'
-    gitUrl = 'https://github.com/aritraroy/PinLockView.git'
-
-    libraryVersion = '2.1.0'
-
-    developerId = 'aritraroy'
-    developerName = 'Aritra Roy'
-    developerEmail = 'aritra.roy.in@gmail.com'
-
-    licenseName = 'The Apache Software License, Version 2.0'
-    licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-    allLicenses = ["Apache-2.0"]
-}
-
+group = 'com.github.msinghal34'
+version = '2.1.0'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    namespace = "com.msinghal34.pinlockview"
+    compileSdk 34
 
     defaultConfig {
-        minSdkVersion 11
-        targetSdkVersion 25
+        minSdkVersion 21
+        targetSdkVersion 34
+
         versionCode 5
         versionName "2.1.0"
+
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+
     }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:recyclerview-v7:25.3.1'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'junit:junit:4.13.2'
+    implementation "androidx.appcompat:appcompat:1.7.0"
+    implementation "androidx.recyclerview:recyclerview:1.3.2"
+    implementation "androidx.annotation:annotation:1.9.0"
 }
 
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
+afterEvaluate {
+    publishing {
+        publications {
+            // Creates a Maven publication called "release".
+            release(MavenPublication) {
+                from components.release
+                groupId = 'com.github.msinghal34'
+                artifactId = 'pinlockview'
+                version = '2.1.0'
+            }
+        }
+
+        repositories {
+            mavenLocal()
+        }
+    }
+}
