@@ -1,11 +1,11 @@
-package com.andrognito.pinlockview;
+package com.msinghal34.pinlockview;
 
 import android.graphics.drawable.Drawable;
 
 /**
  * The customization options for the buttons in {@link PinLockView}
  * passed to the {@link PinLockAdapter} to decorate the individual views
- *
+ * <p>
  * Created by aritraroy on 01/06/16.
  */
 public class CustomizationOptionsBundle {
